pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "orquestra"
version = "0.1.0"
dependencies = [
    "fastapi",
    "uvicorn",
    "sqlalchemy",
    "pydantic",
    "requests",
    "redis",
    "rq",
    "alembic",
    "opentelemetry-api",
    "opentelemetry-sdk",
    "reportlab",
    "jsonschema"
]

[tool.pytest.ini_options]
testpaths = ["tests"]
